Replace instances of EOInt to i32

+++ src/map/map/utils/get_adjacent_tiles.rs
@@ -1,6 +1,6 @@
 use std::cmp;
 
-use eo::{data::i32, protocol::Coords};
+use eolib::protocol::Coords;
 
 use super::super::Map;
 
@@ -10,18 +10,18 @@ impl Map {
         adjacent_tiles.push(*coords);
         adjacent_tiles.push(Coords {
             x: coords.x,
-            y: cmp::max(coords.y as i32 - 1, 0) as i32,
+            y: cmp::max(coords.y - 1, 0) as i32,
         });
         adjacent_tiles.push(Coords {
             x: coords.x,
-            y: cmp::min(coords.y as i32 + 1, self.file.height as i32) as i32,
+            y: cmp::min(coords.y + 1, self.file.height) as i32,
         });
         adjacent_tiles.push(Coords {
-            x: cmp::max(coords.x as i32 - 1, 0) as i32,
+            x: cmp::max(coords.x - 1, 0) as i32,
             y: coords.y,
         });
         adjacent_tiles.push(Coords {
-            x: cmp::min(coords.x as i32 + 1, self.file.width as i32) as i32,
+            x: cmp::min(coords.x + 1, self.file.width) as i32,
             y: coords.y,
         });
